+++ packages/component/src/providers/internal/SendBox/private/types.ts
@@ -1,0 +1,6 @@
+export type ContextType = {
+  submit: (options?: { setFocus?: 'main' | 'sendBox' | 'sendBoxWithoutKeyboard' }) => void;
+  submitErrorMessageIdState: readonly [string];
+};
+
+export type SendError = 'empty' | 'offline';