+++ packages/core/src/sagas/submitSendBoxSaga.js
@@ -10,6 +10,8 @@ function* submitSendBox() {
   yield takeEvery(SUBMIT_SEND_BOX, function* ({ payload: { channelData, method } }) {
     const sendBoxValue = yield select(sendBoxValueSelector);
 
+    // TODO: [P2] If the trimmed value is empty, we should not send.
+    // TODO: [P2] We should expose this logic ("cannot send empty message") to other components such as UI.
     if (sendBoxValue) {
       yield put(sendMessage(sendBoxValue.trim(), method, { channelData }));
       yield put(setSendBox(''));
@@ -18,5 +20,6 @@ function* submitSendBox() {
 }
 
 export default function* submitSendBoxSaga() {
+  // TODO: [P2] We should expose this logic ("send only when connected") to other components such as UI.
   yield whileConnected(submitSendBox);
 }